Move Javascript Vanilla classes to another folder (#36)
