+++ decidim-admin/app/packs/src/decidim/admin/admin_autocomplete.js
@@ -1,4 +1,4 @@
-import AutoComplete from "src/decidim/autocomplete";
+import AutoComplete from "src/decidim/refactor/moved/autocomplete";
 
 /**
  * This function can be used to create an autocomplete input automatically
