+++ decidim-admin/app/packs/src/decidim/admin/participatory_space_search.js
@@ -1,4 +1,4 @@
-import AutoComplete from "src/decidim/autocomplete";
+import AutoComplete from "src/decidim/refactor/moved/autocomplete";
 
 /**
  * Sends a query to the API and resolves the resulting data in the returned
