+++ decidim-core/app/packs/src/decidim/editor/common/upload_dialog.js
@@ -1,4 +1,4 @@
-import icon from "src/decidim/icon";
+import icon from "src/decidim/refactor/moved/icon";
 import { fileNameToTitle } from "src/decidim/editor/utilities/file";
 
 const createElement = (template) => {
