+++ decidim-core/app/packs/src/decidim/refactor/implementation/external_link.js
@@ -1,4 +1,4 @@
-import icon from "src/decidim/icon"
+import icon from "src/decidim/refactor/moved/icon"
 
 const DEFAULT_MESSAGES = {
   externalLink: "External link",
