+++ decidim-core/app/packs/src/decidim/refactor/moved/focus_guard.js
