+++ decidim-core/app/packs/src/decidim/refactor/moved/icon.js
