+++ decidim-core/app/packs/src/decidim/results_listing.js
@@ -1,4 +1,4 @@
-import { pushState, registerCallback } from "src/decidim/history"
+import { pushState, registerCallback } from "src/decidim/refactor/moved/history"
 
 const initializeListingOptionsMenu = (options) => {
   $(document).on("click", `${options.containerSelector} a`, (event) => {
